requirements.txt
numpy
pytesseract
matplotlib
imutils
opencv-python
pillow
pdf2image
Flask
fastapi
python-multipart
albumentations
pandas
setuptools
torch
pyyaml
torchvision
ultralytics
seaborn
scipy
requests
tqdm
psutil
thop
tensorflow
pydantic
gdown
screeninfo
fiftyone
pymupdf
uvicorn
clearml
pycocotools
sahi
EasyOCR
onnx
onnxruntime
